remove zesty explorer
