+++ src/views/zesty/DigitalExperiencePlatform.js
@@ -72,48 +72,1259 @@
  * Images API: https://zesty.org/services/media-storage-micro-dam/on-the-fly-media-optimization-and-dynamic-image-manipulation
  */
 
-import { Box, useTheme } from '@mui/material';
+import {
+  Box,
+  Button,
+  Card,
+  Container,
+  Grid,
+  Link,
+  Typography,
+  useMediaQuery,
+  useTheme,
+} from '@mui/material';
 import React from 'react';
 import SimpleHeroWithImageAndCtaButtonsPage from 'blocks/heroes/SimpleHeroWithImageAndCtaButtons/SimpleHeroWithImageAndCtaButtons.js';
 import FillerContent from 'components/FillerContent';
+import TryFreeButton from 'components/cta/TryFreeButton';
+import * as helper from 'utils';
+import Fade from 'react-reveal/Fade';
+import Zoom from 'react-reveal/Zoom';
+import Slide from '@mui/material/Slide';
+import ArrowRightAltIcon from '@mui/icons-material/ArrowRightAlt';
+import ZoomMui from '@mui/material/Zoom';
+import ZestySvg from '../../../public/assets/images/zesty Logo.png';
+import featuresBg1 from '../../../public/assets/images/dxp_features_bg1.svg';
+import featuresBg2 from '../../../public/assets/images/dxp_features_bg2.svg';
+import headlessCmsBg from '../../../public/assets/images/dxp_headless_bg.svg';
+import bottomBg from '../../../public/assets/images/dxp_bottom_bg.svg';
+import dxpCurve from '../../../public/assets/images/dxp_curve.svg';
+import dxpLine from '../../../public/assets/images/dxp_line.svg';
+import Rotate from 'react-reveal/Rotate';
 
-function DigitalExperiencePlatform({ content }) {
-  console.log(content);
-  const theme = useTheme();
-  //   const headerProps = {
-  //     mainTitle: content.header_title_main || FillerContent.header,
-  //     title: content.header_title || FillerContent.header,
-  //     image:
-  //       (content.header_image?.data && content.header_image?.data[0]?.url) ||
-  //       FillerContent.image,
-  //     description: content.header_description || FillerContent.description,
-  //     cta_left:
-  //       (content.cta_left.data && content.cta_left?.data[0]?.button_text) ||
-  //       FillerContent.cta,
-  //     cta_right:
-  //       (content.cta_right?.data && content.cta_right?.data[0]?.button_text) ||
-  //       FillerContent.cta,
-  //     cta_left_url:
-  //       (content.cta_left?.data &&
-  //         content.cta_left?.data[0]?.internal_link?.data[0]?.meta?.web?.url) ||
-  //       FillerContent.href,
-  //     cta_right_url:
-  //       (content.cta_right.data &&
-  //         content.cta_right?.data[0]?.internal_link?.data[0]?.meta?.web?.url) ||
-  //       FillerContent.href,
-  //   };
-
+const Section1Hero = ({ content, theme, isMobile }) => {
   return (
-    <>
-      {/* <Box
-        position={'relative'}
+    <Box
+      paddingTop={isMobile ? 4 : 15}
+      paddingBottom={isMobile ? 4 : 25}
+      sx={{
+        position: 'relative',
+        background:
+          'linear-gradient(180deg, rgba(31,93,207,1) 45%, rgba(112,152,224,1) 70%, rgba(255,255,255,1) 100%);',
+      }}
+    >
+      <Container
         sx={{
-          backgroundColor: theme.palette.alternate.main,
+          display: 'flex',
+
+          gap: '1rem',
+          flexDirection: isMobile ? 'column' : 'row',
         }}
       >
-        <SimpleHeroWithImageAndCtaButtonsPage />
-      </Box> */}
-    </>
+        <Box
+          sx={{
+            background: '',
+            display: 'flex',
+            justifyContent: 'center',
+            flexDirection: 'column',
+          }}
+        >
+          <Box
+            sx={{
+              position: 'absolute',
+              left: '15rem',
+              display: isMobile ? 'none' : 'flex',
+            }}
+          >
+            <img src={ZestySvg.src} alt="" />
+          </Box>
+          <Typography
+            component={'h2'}
+            variant={'p'}
+            fontWeight={600}
+            sx={{
+              color: theme.palette.common.white,
+              fontWeight: 500,
+            }}
+          >
+            {'DXP'}
+          </Typography>
+          <Typography
+            component={'h1'}
+            variant={'p'}
+            sx={{
+              color: theme.palette.common.white,
+              fontWeight: 'bold',
+              fontSize: isMobile ? '38px' : '48px',
+            }}
+          >
+            {content.header_eyebrow}
+          </Typography>
+          <Typography
+            paddingY={2}
+            component={'h3'}
+            variant={'p'}
+            sx={{
+              color: theme.palette.common.white,
+              fontWeight: 500,
+              fontSize: '20px',
+            }}
+          >
+            {content.header_h1}
+          </Typography>
+          <Box>
+            <Button
+              href={''}
+              component={Button}
+              variant="contained"
+              fullWidth={isMobile ? true : false}
+              sx={{
+                backgroundColor: theme.palette.common.white,
+                color: theme.palette.zesty.zestyOrange,
+                padding: '.6rem 4rem',
+                fontSize: '16px',
+              }}
+            >
+              {content.header_cta_primary}
+            </Button>
+          </Box>
+        </Box>
+
+        <Box
+          sx={{
+            display: 'flex',
+            justifyContent: 'center',
+            justifyItems: 'center',
+          }}
+        >
+          <img
+            src={content.header_graphic.data[0].url}
+            width={isMobile ? 350 : 700}
+          />
+        </Box>
+      </Container>
+    </Box>
+  );
+};
+
+const CustomCard = ({ data, theme }) => {
+  return (
+    <Card
+      sx={{
+        marginTop: '1rem',
+        padding: '3rem 1rem',
+        display: 'flex',
+        alignItems: 'flex-start',
+        borderTop: `6px solid ${theme.palette.zesty.zestyTealWhite}`,
+      }}
+    >
+      <Box paddingX={2} paddingTop={2}>
+        <img src={data.icon} alt="" width={50} />
+      </Box>
+      <Box>
+        <Typography
+          component={'p'}
+          variant={'p'}
+          sx={{
+            color: theme.palette.secondary.darkCharcoal,
+            textAlign: 'left',
+          }}
+          dangerouslySetInnerHTML={{
+            __html: helper.strColorChanger(
+              data.text,
+              data.subText,
+              theme.palette.zesty.zestyOrange,
+            ),
+          }}
+        />
+        <Link
+          href="#"
+          underline="always"
+          paddingTop={4}
+          sx={{
+            display: 'flex',
+            alignItems: 'center',
+            gap: '.5rem',
+            color: theme.palette.zesty.zestyTealDark,
+            fontWeight: 'bold',
+          }}
+        >
+          {data.ctaName} <ArrowRightAltIcon />
+        </Link>
+      </Box>
+    </Card>
+  );
+};
+const SecondSection = ({ content, theme, isMobile }) => {
+  const [active, setactive] = React.useState(0);
+  const headerRegex = /\>(.*?)\</;
+  const cardData = [
+    {
+      id: 0,
+      name: 'personalize',
+      icon: content.solutions_icon_1.data[0].url,
+      text: content.solution_1_description,
+      subText: content.solution_1_description?.match(headerRegex)[1],
+      img: content.solution_1_graphic.data[0].url,
+      ctaName: 'Learn More',
+      href: '#',
+    },
+    {
+      id: 1,
+      name: 'ecom',
+      icon: content.solutions_icon_2.data[0].url,
+      text: content.solution_2_description,
+      subText: content.solution_2_description?.match(headerRegex)[1],
+      img: content.solution_2_graphic.data[0].url,
+      ctaName: 'Learn More',
+      href: '#',
+    },
+    {
+      id: 2,
+      name: 'distribution',
+      icon: content.solutions_icon_3.data[0].url,
+      text: content.solution_3_description,
+      subText: content.solution_3_description?.match(headerRegex)[1],
+      img: content.solution_3_graphic.data[0].url,
+      ctaName: 'Learn More',
+      href: '#',
+    },
+    {
+      id: 3,
+      name: 'innovate',
+      icon: content.solutions_icon_4.data[0].url,
+      text: content.solution_4_description,
+      subText: content.solution_4_description?.match(headerRegex)[1],
+      img: content.solution_4_graphic.data[0].url,
+      ctaName: 'Learn More',
+      href: '#',
+    },
+  ];
+
+  const imgWidth = isMobile ? 300 : 500;
+  return (
+    <Box paddingY={1}>
+      <Container>
+        <Box paddingBottom={isMobile ? 10 : 10} sx={{}}>
+          <Typography
+            component={'h2'}
+            variant={'p'}
+            paddingTop={isMobile ? 0 : 10}
+            sx={{
+              color: theme.palette.secondary.darkCharcoal,
+              fontWeight: 800,
+              fontSize: isMobile ? '20px' : '32px',
+              textAlign: 'center',
+            }}
+            dangerouslySetInnerHTML={{
+              __html: helper.strColorChanger(
+                content.solutions_h2,
+                'digital experience',
+                theme.palette.zesty.zestyOrange,
+              ),
+            }}
+          />
+        </Box>
+        <Box
+          sx={{
+            display: 'flex',
+            gap: isMobile ? '1rem' : '4rem',
+            height: isMobile ? '100vh' : '50vh',
+            justifyContent: 'center',
+            justifyItems: 'center',
+            alignItems: 'center',
+            flexDirection: isMobile ? 'column' : 'row',
+          }}
+        >
+          <Zoom>
+            <Box sx={{}}>
+              <ZoomMui
+                in={active === 0}
+                style={{ display: active === 0 ? 'block' : 'none' }}
+              >
+                <img
+                  src={content[`solution_1_graphic`].data[0].url}
+                  alt=""
+                  width={imgWidth}
+                />
+              </ZoomMui>
+              <ZoomMui
+                in={active === 1}
+                style={{ display: active === 1 ? 'block' : 'none' }}
+              >
+                <img
+                  width={imgWidth}
+                  src={content[`solution_2_graphic`].data[0].url}
+                  alt=""
+                />
+              </ZoomMui>
+              <ZoomMui
+                in={active === 2}
+                style={{ display: active === 2 ? 'block' : 'none' }}
+              >
+                <img
+                  width={imgWidth}
+                  src={content[`solution_3_graphic`].data[0].url}
+                  alt=""
+                />
+              </ZoomMui>
+              <ZoomMui
+                in={active === 3}
+                style={{ display: active === 3 ? 'block' : 'none' }}
+              >
+                <img
+                  width={imgWidth}
+                  src={content[`solution_4_graphic`].data[0].url}
+                  alt=""
+                />
+              </ZoomMui>
+            </Box>
+          </Zoom>
+          <Zoom>
+            <Box sx={{ background: '' }}>
+              {cardData.map((e, i) => {
+                return i === active ? (
+                  <CustomCard data={e} theme={theme} />
+                ) : (
+                  <Box
+                    onClick={() => setactive(i)}
+                    paddingY={2}
+                    paddingLeft={4}
+                    sx={{
+                      borderBottom: `1px solid ${theme.palette.secondary.whiteSmoke}`,
+                      display: 'flex',
+                      alignItems: 'center',
+                      gap: '1rem',
+                      cursor: 'pointer',
+                    }}
+                  >
+                    <img src={e.icon} alt="" width={50} />
+                    <Typography
+                      component={'p'}
+                      variant={'p'}
+                      sx={{
+                        color: theme.palette.secondary.darkCharcoal,
+                        fontWeight: 'light',
+                        textAlign: 'left',
+                      }}
+                      dangerouslySetInnerHTML={{
+                        __html: e.subText,
+                      }}
+                    />
+                  </Box>
+                );
+              })}
+            </Box>
+          </Zoom>
+        </Box>
+        <Zoom>
+          <Grid item xs={12} md={9}>
+            <Box
+              paddingTop={isMobile ? 10 : 10}
+              paddingBottom={10}
+              sx={{
+                textAlign: 'center',
+                color: theme.palette.secondary.darkCharcoal,
+                fontSize: isMobile ? '1rem' : '1.5rem',
+              }}
+              dangerouslySetInnerHTML={{
+                __html: helper.strColorChanger(
+                  content.about_zesty_dxp,
+                  'Zesty',
+                  theme.palette.zesty.zestyOrange,
+                ),
+              }}
+            ></Box>
+          </Grid>
+        </Zoom>
+      </Container>
+    </Box>
+  );
+};
+
+const ThirdSection = ({ content, theme, isMobile }) => {
+  return (
+    <Box
+      sx={{
+        background: theme.palette.zesty.zestySeaShell,
+      }}
+      paddingBottom={isMobile ? 1 : 20}
+      paddingTop={isMobile ? 1 : 10}
+    >
+      <Container>
+        <Box>
+          <Typography
+            component={'p'}
+            variant={'p'}
+            paddingBottom={isMobile ? 2 : 10}
+            sx={{
+              fontSize: isMobile ? '1rem' : '1.4rem',
+              color: theme.palette.secondary.darkCharcoal,
+              textAlign: 'center',
+            }}
+            dangerouslySetInnerHTML={{
+              __html: helper.strColorChanger(
+                content.about_dxp,
+                'Digital Experience Platform?',
+                theme.palette.zesty.zestyOrange,
+              ),
+            }}
+          />
+        </Box>
+
+        <Box
+          sx={{
+            display: 'flex',
+            justifyItems: 'center',
+            justifyContent: 'center',
+            position: 'relative',
+            zIndex: '1000',
+          }}
+        >
+          <Zoom>
+            <img
+              src={content.about_dxp_graphic.data[0].url}
+              alt=""
+              width={isMobile ? 300 : 700}
+            />
+          </Zoom>
+        </Box>
+      </Container>
+    </Box>
+  );
+};
+
+const RevealComponent = ({
+  text = '',
+  img = FillerContent.dashboard_image,
+  strToChange = '',
+  reverse = false,
+  theme,
+  strColor = theme.palette.zesty.zestyOrange,
+  data,
+  index,
+  isMobile,
+}) => {
+  return (
+    <Box
+      paddingY={1}
+      sx={{
+        display: 'flex',
+        flexDirection: isMobile ? 'column' : reverse ? 'row-reverse' : 'row',
+        position: 'relative',
+        zIndex: '1000',
+      }}
+    >
+      {index !== 3 && (
+        <Box
+          sx={{
+            zIndex: '10',
+            background: theme.palette.zesty.zestyBackgroundBlue,
+            height: isMobile ? '15px' : '20px',
+            width: isMobile ? '15px' : '20px',
+            borderRadius: '50%',
+            position: 'absolute',
+            left: isMobile ? '1.8%' : '50.2%',
+            bottom: isMobile ? '45%' : '0',
+            transform: 'translate(-50%,-50%)',
+          }}
+        />
+      )}
+
+      <Fade left>
+        <Box sx={{ position: 'relative' }}>
+          <Typography
+            component={'h3'}
+            variant={'p'}
+            paddingLeft={isMobile ? 2 : 0}
+            paddingTop={isMobile ? 2 : 10}
+            paddingBottom={isMobile ? 6 : 10}
+            sx={{
+              color: theme.palette.common.white,
+              textAlign: 'left',
+              fontSize: isMobile ? '1rem' : '1.5rem',
+            }}
+            dangerouslySetInnerHTML={{
+              __html: helper.strColorChanger(text, strToChange, strColor),
+            }}
+          />
+          <Link
+            href="#"
+            underline="always"
+            paddingLeft={isMobile ? 2 : 0}
+            sx={{
+              position: 'absolute',
+              top: isMobile ? '11rem' : '22rem',
+              display: 'flex',
+              alignItems: 'center',
+              gap: '.5rem',
+              color: theme.palette.zesty.zestyTealDark,
+              fontWeight: 'bold',
+            }}
+          >
+            Learn More <ArrowRightAltIcon />
+          </Link>
+        </Box>
+      </Fade>
+      <Fade right>
+        <img src={img} width={isMobile ? 350 : 700} />
+      </Fade>
+    </Box>
+  );
+};
+
+const FourthSection = ({ content, theme, isMobile }) => {
+  const FillerImage =
+    content.middle_solution_1_graphic.data[0].url ||
+    FillerContent.dashboard_image;
+  const arr = [
+    {
+      text: content.middle_solution_1_description,
+      img:
+        content.middle_solution_1_graphic &&
+        content.middle_solution_1_graphic.data[0].url,
+      strToChange: 'Personalized experience',
+      reverse: false,
+      theme: theme,
+    },
+    {
+      text: content.middle_solution_2_description,
+      img:
+        (content.middle_solution_2_graphic &&
+          content.middle_solution_2_graphic?.data[0]?.url) ||
+        FillerImage,
+      strToChange: 'Commerce experience',
+      reverse: true,
+      theme: theme,
+    },
+    {
+      text: content.middle_solution_3_description,
+      img:
+        (content.middle_solution_3_graphic &&
+          content.middle_solution_3_graphic.data[0]?.url) ||
+        FillerImage,
+      strToChange: 'Content Experience',
+      reverse: false,
+      theme: theme,
+    },
+    {
+      text: content.middle_solution_4_description,
+      img:
+        (content.middle_solution_4_graphic &&
+          content.middle_solution_4_graphic.data[0]?.url) ||
+        FillerImage,
+      strToChange: 'Enterprise Experience',
+      reverse: true,
+      theme: theme,
+    },
+  ];
+  return (
+    <Box
+      paddingY={isMobile ? 2 : 10}
+      sx={{
+        backgroundColor: theme.palette.zesty.zestyDarkBlue,
+        position: 'relative',
+      }}
+    >
+      <Box
+        sx={{
+          zIndex: '10',
+          position: 'absolute',
+          left: '5%',
+          top: 0,
+          display: !isMobile ? 'none' : 'flex',
+        }}
+      >
+        <img src={dxpCurve.src} alt="bg" />
+      </Box>
+      <Box
+        sx={{
+          zIndex: '10',
+          position: 'absolute',
+          left: '5%',
+          bottom: 0,
+          display: !isMobile ? 'none' : 'flex',
+        }}
+      >
+        <img src={dxpLine.src} alt="bg" />
+      </Box>
+      <Box
+        sx={{
+          zIndex: '10',
+          position: 'absolute',
+          left: 0,
+          bottom: 0,
+          display: isMobile ? 'none' : 'flex',
+        }}
+      >
+        <img src={featuresBg1.src} alt="bg" />
+      </Box>
+      <Box
+        sx={{
+          zIndex: '10',
+          position: 'absolute',
+          right: 0,
+          top: 0,
+          display: isMobile ? 'none' : 'flex',
+        }}
+      >
+        <img src={featuresBg2.src} alt="bg" />
+      </Box>
+      <Box
+        sx={{
+          zIndex: '10',
+          background: theme.palette.zesty.zestyBackgroundBlue,
+          height: isMobile ? '0' : '260vh',
+          width: '2px',
+          position: 'absolute',
+          left: isMobile ? '5%' : '50%',
+          bottom: '0',
+        }}
+      />
+      <Container>
+        <Box>
+          <Typography
+            component={'h2'}
+            variant={'p'}
+            paddingTop={isMobile ? 8 : 10}
+            paddingBottom={isMobile ? 0 : 10}
+            sx={{
+              color: theme.palette.common.white,
+              textAlign: 'center',
+              fontSize: isMobile ? '28px' : '48px',
+              position: 'relative',
+              zIndex: '1000',
+              textTransform: 'capitalize',
+            }}
+          >
+            {content.middle_solutions_header}
+          </Typography>
+        </Box>
+        {arr.map((e, i) => {
+          return (
+            <RevealComponent
+              isMobile={isMobile}
+              data={e}
+              index={i}
+              text={e.text}
+              img={e.img}
+              strToChange={e.strToChange}
+              reverse={e.reverse}
+              theme={e.theme}
+            />
+          );
+        })}
+      </Container>
+    </Box>
+  );
+};
+const FifthSection = ({ content, theme, isMobile }) => {
+  const arr = content.features.data;
+  return (
+    <Box paddingBottom={20} sx={{ position: 'relative' }}>
+      <Box
+        sx={{
+          zIndex: '10',
+          position: 'absolute',
+          left: 0,
+          bottom: 0,
+          display: isMobile ? 'none' : 'flex',
+        }}
+      >
+        <img src={featuresBg1.src} alt="bg" />
+      </Box>
+      <Container>
+        <Box>
+          <Typography
+            component={'h2'}
+            variant={'p'}
+            paddingTop={isMobile ? 4 : 10}
+            paddingBottom={isMobile ? 4 : 10}
+            sx={{
+              color: theme.palette.secondary.darkCharcoal,
+              textAlign: 'center',
+              fontSize: isMobile ? '24px' : '48px',
+            }}
+            dangerouslySetInnerHTML={{
+              __html: helper.strColorChanger(
+                content.features_header,
+                'Zesty',
+                theme.palette.zesty.zestyOrange,
+              ),
+            }}
+          />
+        </Box>
+        <Box
+          sx={{
+            display: 'flex',
+            flexWrap: 'wrap',
+            justifyItems: 'center',
+            justifyContent: 'center',
+            gap: '4rem',
+            position: 'relative',
+            zIndex: '1000',
+          }}
+        >
+          {arr?.map((e) => {
+            return (
+              <Zoom>
+                <Card
+                  sx={{
+                    width: '20rem',
+                    height: '22rem',
+                    padding: '3rem 2rem',
+                    background: theme.palette.common.white,
+                  }}
+                >
+                  <img src={e.icon_image.data[0].url} alt="" />
+
+                  <Typography
+                    component={'p'}
+                    variant={'p'}
+                    paddingTop={4}
+                    paddingBottom={2}
+                    sx={{
+                      color: theme.palette.zesty.zestyOrange,
+                      textAlign: 'left',
+                      fontSize: '20px',
+                    }}
+                  >
+                    {e?.feature_name}
+                  </Typography>
+                  <Typography
+                    component={'h2'}
+                    variant={'p'}
+                    sx={{
+                      color: theme.palette.common.dark,
+                      textAlign: 'left',
+                      fontSize: '16px',
+                      fontWeight: 'light',
+                    }}
+                  >
+                    {e?.content}
+                  </Typography>
+                </Card>
+              </Zoom>
+            );
+          })}
+        </Box>
+      </Container>
+    </Box>
+  );
+};
+
+const SixtheSection = ({ content, theme, isMobile }) => {
+  return (
+    <Box
+      paddingY={2}
+      sx={{
+        background: theme.palette.zesty.zestyGray99,
+      }}
+    >
+      <Container
+        sx={{
+          display: 'flex',
+          alignItems: 'center',
+          flexDirection: isMobile ? 'column-reverse' : 'row',
+          gap: '4rem',
+        }}
+      >
+        <Zoom>
+          <Box sx={{}}>
+            <img
+              src={content.integrations_graphic.data[0].url}
+              width={isMobile ? 400 : 700}
+            />
+          </Box>
+        </Zoom>
+        <Box
+          sx={{
+            background: '',
+            display: 'flex',
+            justifyContent: 'center',
+            flexDirection: 'column',
+            position: 'relative',
+          }}
+        >
+          <Rotate right>
+            <Box sx={{ position: 'absolute', top: '-5rem' }}>
+              <img
+                src={content.integrations_airplane_graphic.data[0].url}
+                alt=""
+              />
+            </Box>
+          </Rotate>
+          <Fade right>
+            <Typography
+              component={'p'}
+              variant={'p'}
+              paddingY={isMobile ? 6 : 10}
+              sx={{
+                color: theme.palette.secondary.darkCharcoal,
+                textAlign: 'left',
+              }}
+              dangerouslySetInnerHTML={{
+                __html: helper.strColorChanger(
+                  content.integrations_description,
+                  'Limitless Integrations for Limitless Reach',
+                  theme.palette.zesty.zestyOrange,
+                ),
+              }}
+            />
+            <Box>
+              <Button
+                variant="contained"
+                size="large"
+                fullWidth={isMobile}
+                // href={
+                //   content.integrations_button_link?.data[0]?.meta?.web?.uri ||
+                //   ''
+                // }
+                sx={{
+                  backgroundColor: theme.palette.zesty.zestyOrange,
+                  color: theme.palette.common.white,
+                  padding: '.6rem 4rem',
+                  fontSize: '16px',
+                }}
+              >
+                {content.integrations_button_text}
+              </Button>
+            </Box>
+          </Fade>
+        </Box>
+      </Container>
+    </Box>
+  );
+};
+
+const Section7 = ({ content, theme, isMobile }) => {
+  const [headless, setheadless] = React.useState(true);
+  const [hybrid, sethybrid] = React.useState(false);
+  return (
+    <Box
+      paddingY={isMobile ? 4 : 15}
+      sx={{
+        background: theme.palette.common.white,
+        position: 'relative',
+      }}
+    >
+      <Box
+        sx={{
+          zIndex: '10',
+          position: 'absolute',
+          right: 0,
+          top: 0,
+          display: isMobile ? 'none' : 'flex',
+        }}
+      >
+        <img src={headlessCmsBg.src} alt="bg" />
+      </Box>
+      <Container
+        sx={{
+          alignItems: 'center',
+        }}
+      >
+        <Box
+          sx={{
+            display: 'flex',
+            textAlign: 'center',
+            justifyItems: 'center',
+            justifyContent: 'center',
+          }}
+        >
+          <Typography
+            component={'h2'}
+            variant={'p'}
+            paddingBottom={8}
+            sx={{
+              fontSize: isMobile ? '22px' : '32px',
+              color: theme.palette.secondary.darkCharcoal,
+              textAlign: 'center',
+            }}
+            dangerouslySetInnerHTML={{
+              __html: helper.strColorChanger(
+                content.implementing_header,
+                'Digital Experience',
+                theme.palette.zesty.zestyOrange,
+              ),
+            }}
+          />
+        </Box>
+        <Box
+          sx={{
+            width: '100%',
+            display: 'flex',
+            justifyItems: 'center',
+            justifyContent: 'center',
+          }}
+        >
+          <Card sx={{ display: 'flex' }}>
+            <Box
+              onClick={() => {
+                setheadless(true);
+                sethybrid(false);
+              }}
+              sx={{
+                display: 'flex',
+                alignItems: 'center',
+                padding: isMobile ? '1rem 1rem' : '1rem 5rem',
+                gap: '1rem',
+                position: 'relative',
+                zIndex: '1000',
+                cursor: 'pointer',
+                color: headless
+                  ? theme.palette.zesty.zestyTealDark
+                  : theme.palette.common.dark,
+                borderBottom: `4px solid ${
+                  headless ? theme.palette.zesty.zestyTealDark : 'transparent'
+                }`,
+              }}
+            >
+              <img
+                src={content.headless_cms_toggle_graphic.data[0].url}
+                alt=""
+              />
+              <Typography
+                component={'p'}
+                variant={'p'}
+                sx={{
+                  textAlign: 'left',
+                  color: headless
+                    ? theme.palette.zesty.zestyTealDark
+                    : theme.palette.secondary.darkCharcoal,
+                }}
+              >
+                {content.headless_cms_toggle}
+              </Typography>
+            </Box>
+            <Box
+              onClick={() => {
+                setheadless(false);
+                sethybrid(true);
+              }}
+              sx={{
+                display: 'flex',
+                cursor: 'pointer',
+                alignItems: 'center',
+                padding: isMobile ? '1rem 1rem' : '1rem 5rem',
+                position: 'relative',
+                zIndex: '1000',
+                gap: '1rem',
+                borderBottom: `4px solid ${
+                  hybrid ? theme.palette.zesty.zestyTealDark : 'transparent'
+                }`,
+              }}
+            >
+              <img alt="" src={content.hybrid_cms_toggle_graphic.data[0].url} />
+              <Typography
+                component={'p'}
+                variant={'p'}
+                sx={{
+                  textAlign: 'left',
+                  color: hybrid
+                    ? theme.palette.zesty.zestyTealDark
+                    : theme.palette.secondary.darkCharcoal,
+                }}
+              >
+                {content.hybrid_cms_toggle}
+              </Typography>
+            </Box>
+          </Card>{' '}
+        </Box>
+        <Box
+          paddingTop={isMobile ? 4 : 15}
+          sx={{
+            display: 'flex',
+            position: 'relative',
+            zIndex: '1000',
+            flexDirection: isMobile ? 'column' : 'row',
+            justifyContent: 'center',
+            justifyItems: 'center',
+          }}
+        >
+          <Zoom>
+            <ZoomMui
+              in={headless}
+              style={{
+                display: headless ? 'flex' : 'none',
+                justifyContent: 'center',
+                justifyItems: 'center',
+                alignItems: 'center',
+                margin: '0 auto',
+              }}
+            >
+              <img
+                src={content.headless_cms_graphic.data[0].url}
+                width={isMobile ? 350 : 600}
+              />
+            </ZoomMui>
+            <ZoomMui
+              in={hybrid}
+              style={{
+                display: hybrid ? 'flex' : 'none',
+                justifyContent: 'center',
+                justifyItems: 'center',
+                alignItems: 'center',
+
+                margin: '0 auto',
+              }}
+            >
+              <img
+                width={isMobile ? 350 : 600}
+                src={content.headless_cms_graphic.data[0].url}
+              />
+            </ZoomMui>
+          </Zoom>
+
+          <Zoom>
+            <ZoomMui
+              in={headless}
+              style={{ display: headless ? 'block' : 'none' }}
+            >
+              <Box
+                sx={{
+                  display: 'flex',
+                  justifyContent: 'center',
+                  flexDirection: 'column',
+                  position: 'relative',
+                }}
+              >
+                <Typography
+                  component={'p'}
+                  variant={'p'}
+                  paddingY={isMobile ? 4 : 10}
+                  sx={{
+                    color: theme.palette.secondary.darkCharcoal,
+                    textAlign: 'left',
+                    fontSize: '1.2rem',
+                  }}
+                  dangerouslySetInnerHTML={{
+                    __html: helper.strColorChanger(
+                      content.headless_cms_description,
+                      'Personalize at scale with Data',
+                      theme.palette.zesty.zestyOrange,
+                    ),
+                  }}
+                />
+                <Link
+                  href="#"
+                  underline="always"
+                  sx={{
+                    position: 'absolute',
+                    top: isMobile ? '10rem' : '10rem',
+                    display: 'flex',
+                    alignItems: 'center',
+                    gap: '.5rem',
+                    color: theme.palette.zesty.zestyTealDark,
+                  }}
+                >
+                  Learn More <ArrowRightAltIcon />
+                </Link>
+              </Box>
+            </ZoomMui>
+            <ZoomMui in={hybrid} style={{ display: hybrid ? 'block' : 'none' }}>
+              <Box
+                sx={{
+                  display: 'flex',
+                  justifyContent: 'center',
+                  flexDirection: 'column',
+                  position: 'relative',
+                }}
+              >
+                <Typography
+                  component={'p'}
+                  variant={'p'}
+                  paddingY={isMobile ? 4 : 10}
+                  sx={{
+                    color: theme.palette.secondary.darkCharcoal,
+                    textAlign: 'left',
+                    fontSize: '1.2rem',
+                  }}
+                  dangerouslySetInnerHTML={{
+                    __html: helper.strColorChanger(
+                      content.hybrid_cms_description,
+                      'Personalize at scale with Data',
+                      theme.palette.zesty.zestyOrange,
+                    ),
+                  }}
+                />
+                <Link
+                  href="#"
+                  underline="always"
+                  sx={{
+                    position: 'absolute',
+                    top: '25vh',
+                    display: 'flex',
+                    alignItems: 'center',
+                    gap: '.5rem',
+                    color: theme.palette.zesty.zestyTealDark,
+                  }}
+                >
+                  Learn More <ArrowRightAltIcon />
+                </Link>
+              </Box>
+            </ZoomMui>
+          </Zoom>
+        </Box>
+      </Container>
+    </Box>
+  );
+};
+
+const Section8CaseStudies = ({ content, theme }) => {
+  return (
+    <Box>
+      <Container>
+        <Typography
+          component={'h2'}
+          variant={'p'}
+          paddingBottom={10}
+          sx={{
+            fontSize: '32px',
+            color: theme.palette.secondary.darkCharcoal,
+            textAlign: 'center',
+          }}
+          dangerouslySetInnerHTML={{
+            __html: helper.strColorChanger(
+              content.case_study_header,
+              '',
+              theme.palette.zesty.zestyOrange,
+            ),
+          }}
+        />
+      </Container>
+      <Box
+        sx={{ height: '30rem', background: theme.palette.zesty.zestyDarkBlue }}
+      >
+        <img src="" />
+      </Box>
+      <Container>
+        <Box
+          paddingY={4}
+          sx={{
+            display: 'flex',
+            flexWrap: 'wrap',
+            justifyContent: 'center',
+            justifyItems: 'center',
+            gap: '3rem',
+          }}
+        >
+          {content.case_studies?.data?.map((e) => {
+            return <img src={e.logo.data[0].url} />;
+          })}
+        </Box>
+      </Container>
+    </Box>
+  );
+};
+const Section9Bottom = ({ content, theme, isMobile }) => {
+  return (
+    <Box paddingY={isMobile ? 0 : 20} sx={{ position: 'relative' }}>
+      <Box
+        sx={{
+          zIndex: '10',
+          position: 'absolute',
+          right: 0,
+          bottom: 0,
+          display: isMobile ? 'none' : 'flex',
+        }}
+      >
+        <img src={bottomBg.src} alt="bg" />
+      </Box>
+      <Container>
+        <Box
+          paddingY={isMobile ? 2 : 8}
+          sx={{
+            display: 'flex',
+            flexDirection: isMobile ? 'column' : 'row-reverse',
+            gap: isMobile ? '1rem' : '4rem',
+          }}
+        >
+          <Fade left>
+            <Box sx={{ position: 'relative' }}>
+              <Typography
+                component={'h2'}
+                variant={'p'}
+                sx={{
+                  fontSize: isMobile ? '.9rem' : '1.3rem',
+                  color: theme.palette.secondary.darkCharcoal,
+                  textAlign: isMobile ? 'center' : 'left',
+                }}
+                dangerouslySetInnerHTML={{
+                  __html: helper.strColorChanger(
+                    content.bottom_cta_description,
+                    'Digital Experiences',
+                    theme.palette.zesty.zestyOrange,
+                  ),
+                }}
+              />
+              <Box
+                sx={{
+                  position: 'absolute',
+                  top: isMobile ? '6rem' : '10rem',
+                  display: 'flex',
+                  flexDirection: isMobile ? 'column' : 'row',
+                  alignItems: 'center',
+                  gap: '.5rem',
+                  color: theme.palette.zesty.zestyTealDark,
+                  width: '100%',
+                }}
+              >
+                <Button
+                  sx={{ padding: '.5rem 4rem' }}
+                  variant="contained"
+                  color="secondary"
+                  fullWidth={isMobile}
+                >
+                  {content.bottom_cta_button_primary}
+                </Button>
+                <Button
+                  variant="text"
+                  color="secondary"
+                  fullWidth={isMobile}
+                  sx={{ textDecoration: 'underline' }}
+                >
+                  {content.bottom_cta_button_secondary}
+                  <ArrowRightAltIcon />
+                </Button>
+              </Box>
+            </Box>
+          </Fade>
+          <Fade right>
+            <Box paddingY={isMobile ? 15 : 0}>
+              <img
+                src={content.bottom_cta_graphic.data[0].url}
+                width={isMobile ? 350 : 600}
+              />
+            </Box>
+          </Fade>
+        </Box>
+      </Container>
+    </Box>
+  );
+};
+
+function DigitalExperiencePlatform({ content }) {
+  console.log(content, 12222);
+  const theme = useTheme();
+  const isMobile = useMediaQuery(theme.breakpoints.down('sm'));
+
+  return (
+    <Box sx={{ overflowX: 'hidden' }}>
+      <Section1Hero content={content} theme={theme} isMobile={isMobile} />
+      <SecondSection content={content} theme={theme} isMobile={isMobile} />
+      <ThirdSection content={content} theme={theme} isMobile={isMobile} />
+      <FourthSection content={content} theme={theme} isMobile={isMobile} />
+      <FifthSection content={content} theme={theme} isMobile={isMobile} />
+      <SixtheSection content={content} theme={theme} isMobile={isMobile} />
+      <Section7 content={content} theme={theme} isMobile={isMobile} />
+      <Section8CaseStudies
+        content={content}
+        theme={theme}
+        isMobile={isMobile}
+      />
+      <Section9Bottom content={content} theme={theme} isMobile={isMobile} />
+    </Box>
   );
 }
